+++ NEWS.md
@@ -1,7 +1,7 @@
 # hutils 1.0.0
 * Stable release:
   - All tests are now marked as major release.
-
+* New function `%notchin%` for a 'safer' alternative to `%notin%`.
 
 # hutils 0.10.0
 * New function `implies`, logical implies.